fix: [CHK-2502] split event processing in separate class
