+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsingestor/event/EventHandler.java
@@ -4,6 +4,7 @@ import it.gov.pagopa.rtd.ms.rtdmsingestor.model.BlobApplicationAware;
 import it.gov.pagopa.rtd.ms.rtdmsingestor.model.BlobApplicationAware.Status;
 import it.gov.pagopa.rtd.ms.rtdmsingestor.model.EventGridEvent;
 import it.gov.pagopa.rtd.ms.rtdmsingestor.service.BlobRestConnector;
+import it.gov.pagopa.rtd.ms.rtdmsingestor.service.EventProcessor;
 import java.util.List;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
@@ -26,12 +27,12 @@ public class EventHandler {
    */
   @Bean
   public Consumer<Message<List<EventGridEvent>>> blobStorageConsumer(
-      BlobRestConnector blobRestConnector) {
+      BlobRestConnector blobRestConnector, EventProcessor blobProcessor) {
     return message -> message.getPayload().stream()
         .filter(e -> "Microsoft.Storage.BlobCreated".equals(e.getEventType()))
         .map(EventGridEvent::getSubject).map(BlobApplicationAware::new)
         .filter(b -> Status.RECEIVED.equals(b.getStatus())).map(blobRestConnector::get)
-        .filter(b -> Status.DOWNLOADED.equals(b.getStatus())).map(blobRestConnector::process)
+        .filter(b -> Status.DOWNLOADED.equals(b.getStatus())).map(blobProcessor::process)
         .filter(b -> Status.PROCESSED.equals(b.getStatus())).map(blobRestConnector::deleteRemote)
         .filter(b -> Status.REMOTELY_DELETED.equals(b.getStatus()))
         .map(BlobApplicationAware::localCleanup)
