+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsingestor/service/BlobRestConnector.java
@@ -43,7 +43,7 @@ import org.springframework.validation.annotation.Validated;
 @Slf4j
 @Validated
 @RequiredArgsConstructor
-public class BlobRestConnector implements TransactionCheck {
+public class BlobRestConnector {
 
   @Value("${ingestor.api.baseurl}")
   private String baseUrl;
@@ -56,13 +56,6 @@ public class BlobRestConnector implements TransactionCheck {
 
   private final CloseableHttpClient httpClient;
 
-  private final StreamBridge sb;
-
-  private final IngestorRepository repository;
-
-  private int numNotEnrolledCards = 0;
-  private int numCorrectTrx = 0;
-  private int numTotalTrx = 0;
 
   /**
    * Method that allows the get of the blob from a remote storage.
@@ -86,65 +79,6 @@ public class BlobRestConnector implements TransactionCheck {
           ex.getMessage());
     }
 
-    return blob;
-  }
-
-  /**
-   * Method that maps transaction fields taken them from csv into Transaction object, then send it
-   * on the output queue. This is done for each transaction inside the blob received.
-   *
-   * @param blob the blob of the transaction.
-   */
-  public BlobApplicationAware process(BlobApplicationAware blob) {
-    log.info("Extracting transactions from:{}", blob.getBlobUri());
-
-    FileReader fileReader;
-
-    numTotalTrx = 0;
-    numCorrectTrx = 0;
-    numNotEnrolledCards = 0;
-
-    try {
-      fileReader = new FileReader(Path.of(blob.getTargetDir(), blob.getBlob()).toFile());
-    } catch (FileNotFoundException e) {
-      log.error("Missing blob file: {}", blob.getBlob());
-      return blob;
-    }
-
-    BeanVerifier<Transaction> verifier = new TransactionVerifier();
-
-    CsvToBeanBuilder<Transaction> builder =
-        new CsvToBeanBuilder<Transaction>(fileReader).withType(Transaction.class).withSeparator(';')
-            .withVerifier(verifier).withThrowExceptions(false);
-
-    CsvToBean<Transaction> csvToBean = builder.build();
-    Stream<Transaction> readTransaction = csvToBean.stream();
-
-    transactionCheckProcess(readTransaction);
-
-    List<CsvException> violations = csvToBean.getCapturedExceptions();
-
-    numTotalTrx = numTotalTrx + violations.size();
-
-    if (!violations.isEmpty()) {
-      for (CsvException e : violations) {
-        log.error("Validation error at line " + e.getLineNumber() + " : " + e.getMessage());
-      }
-    } else if (numTotalTrx == 0) {
-      log.error("No records found in file {}", blob.getBlob());
-    }
-
-    if (numTotalTrx == numCorrectTrx) {
-      log.info("Extraction result: extracted all {} transactions from:{}", numCorrectTrx,
-          blob.getBlobUri());
-    } else {
-      log.info(
-          "Extraction result: {} well formed transactions and {} "
-              + "not enrolled cards out of {} rows extracted from:{}",
-          numCorrectTrx, numNotEnrolledCards, numTotalTrx, blob.getBlobUri());
-    }
-
-    blob.setStatus(Status.PROCESSED);
     return blob;
   }
 
@@ -191,34 +125,6 @@ public class BlobRestConnector implements TransactionCheck {
     }
   }
 
-  protected int getNumNotEnrolledCards() {
-    return numNotEnrolledCards;
-  }
 
-  protected int getNumTotalTrx() {
-    return numTotalTrx;
-  }
 
-  protected int getNumCorrectTrx() {
-    return numCorrectTrx;
-  }
-
-  @Override
-  public void transactionCheckProcess(Stream<Transaction> readTransaction) {
-    readTransaction.forEach(t -> {
-      try {
-        Optional<EPIItem> dbResponse = repository.findItemByHash(t.getHpan());
-        if (dbResponse.isPresent()) {
-          t.setHpan(dbResponse.get().getHashPan());
-          sb.send("rtdTrxProducer-out-0", MessageBuilder.withPayload(t).build());
-          numCorrectTrx++;
-        } else {
-          numNotEnrolledCards++;
-        }
-        numTotalTrx++;
-      } catch (Exception ex) {
-        log.error("Error getting records : {}", ex.getMessage());
-      }
-    });
-  }
 }
